requirements.txt
pillow
requests
tqdm